zip: support Linux device metadata on 32-bit architectures

+++ sys_linux.go
@@ -18,11 +18,11 @@ func sysPlatformExtra(fi os.FileInfo, hdr *FileHeader) {
 
 	switch fi.Mode() & os.ModeType {
 	case os.ModeDevice | os.ModeCharDevice:
-		hdr.Devmajor = int64(unix.Major(sys.Rdev))
-		hdr.Devminor = int64(unix.Minor(sys.Rdev))
+		hdr.Devmajor = int64(unix.Major(uint64(sys.Rdev)))
+		hdr.Devminor = int64(unix.Minor(uint64(sys.Rdev)))
 	case os.ModeDevice:
-		hdr.Devmajor = int64(unix.Major(sys.Rdev))
-		hdr.Devminor = int64(unix.Minor(sys.Rdev))
+		hdr.Devmajor = int64(unix.Major(uint64(sys.Rdev)))
+		hdr.Devminor = int64(unix.Minor(uint64(sys.Rdev)))
 	}
 }
